% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_Variance.R
\name{func.boot}
\alias{func.boot}
\title{func.boot}
\usage{
func.boot(data, indices)
}
\arguments{
\item{data}{matrix with indices of possible imputed datasets to sample}

\item{indices}{sampled indices}
}
\description{
This function is called internal to Bootstrap_Variance and re-estimates glm model parameters
}
