% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Louis_Information.R
\name{Louis_Information}
\alias{Louis_Information}
\title{Louis_Information}
\usage{
Louis_Information(fit, stack, M, IMPUTED)
}
\arguments{
\item{fit}{object of class glm or coxph from fitting to the stacked dataset}

\item{stack}{data frame containing stacked dataset across multiple imputations. Could have 1 or M rows for each subject with complete data. Should have M rows for each subject with imputed data. Must contain the following named columns: (1) stack$.id, which correspond to a unique identifier for each subject. This column can be easily output from MICE. (2) stack$wt, which corresponds to weights assigned to each row. Standard analysis of stacked multiple imputations should set these weights to 1 over the number of times the subject appears in the stack.}

\item{M}{number of multiple imputations}

\item{IMPUTED}{vector of subject ids (matching stack$.id) corresponding to subjects with imputed data. Could specify this vector to include all patient ids, but this will result in slower estimation.}
}
\value{
Info, estimated information matrix accounting for within and between imputation variation
}
\description{
This function takes a dataset with stacked multiple imputations and a glm or coxph fit and estimates corresponding standard errors.
}
\details{
This function uses a variant of the observed information matrix estimator proposed in Louis (1982) and extended in Wei and Tanner (1990). This estimator is a further extension pecifically designed for analyzing stacks of multiply imputed data.
}
