% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_Variance.R
\name{my_update}
\alias{my_update}
\title{my_update}
\usage{
my_update(mod, formula = NULL, data = NULL, weights = NULL)
}
\arguments{
\item{mod}{object of class 'glm' or 'coxph'}

\item{formula}{formula for updated model fit, default = no change}

\item{data}{data used for updated model fit, default = no change}

\item{weights}{weights used for updated model fit, default = no change}
}
\value{
the updated model fit object of the same class as the given model
}
\description{
Function for updating a model fit using either new data or a new model structure
}
