% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stackExample}
\alias{stackExample}
\title{Example data for Louis_Information()}
\format{
a list with
\itemize{
\item{fit}{ glm fit from vignette example}
\item{stack}{ stacked imputed data sets from vignette example}
}
}
\description{
Example data set for Louis_Information()
}
\keyword{data}
