% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_Variance.R
\name{my_update}
\alias{my_update}
\title{Title}
\usage{
my_update(mod, formula = NULL, data = NULL, weights = NULL)
}
\arguments{
\item{mod}{desc}

\item{formula}{desc}

\item{data}{desc}

\item{weights}{desc}
}
\description{
Description
}
